+++ guillotina/db/transaction_manager.py
@@ -138,7 +138,9 @@ class TransactionManager:
                         raise
             except Exception:
                 # failsafe terminate to make sure connection is cleaned
-                if txn._db_conn is not None:
+                if txn._db_conn is None:
+                    raise
+                if txn._db_conn._con is None:
                     raise
                 try:
                     await self._storage.terminate(txn._db_conn)