+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ CHANGELOG
 5.0.18 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix connection leaks in edge-cases
+  [masipcat]
+
+- Pinned asyncpg to 0.19.0
+  [masipcat]
 
 
 5.0.17 (2019-10-11)